Renamed filenames and changes the routes and relative paths. At least most of them.....
