+++ db/connection.js
@@ -2,11 +2,11 @@
 const { Pool } = require('pg');
 
 const dbParams = {
-  host: process.env.DB_HOST,
-  port: process.env.DB_PORT,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASS,
-  database: process.env.DB_NAME
+  host: 'localhost',
+  port: 5432,
+  user: 'labber',
+  password: 'labber',
+  database: 'midterm'
 };
 
 const db = new Pool(dbParams);
