+++ db/queries/product_query.js
@@ -1,0 +1,13 @@
+// what type of queries do we need from product?
+//
+
+const db = require('../connection');
+
+const getProducts = () => {
+  return db.query('SELECT * FROM products;')
+    .then(data => {
+      return data.rows;
+    });
+};
+
+module.exports = { getProducts };
