+++ routes/users-api.js
@@ -7,7 +7,7 @@
 
 const express = require('express');
 const router  = express.Router();
-const userQueries = require('../db/queries/users');
+const userQueries = require('../db/queries/users_query');
 
 router.get('/', (req, res) => {
   userQueries.getUsers()
