+++ server.js
@@ -29,14 +29,14 @@ app.use(express.static('public'));
 // Separated Routes for each Resource
 // Note: Feel free to replace the example routes below with your own
 const userApiRoutes = require('./routes/users-api');
-const widgetApiRoutes = require('./routes/widgets-api');
-const usersRoutes = require('./routes/users');
+const productApiRoutes = require('./routes/products-api');
+const usersRoutes = require('./routes/users_routes');
 
 // Mount all resource routes
 // Note: Feel free to replace the example routes below with your own
 // Note: Endpoints that return data (eg. JSON) usually start with `/api`
 app.use('/api/users', userApiRoutes);
-app.use('/api/widgets', widgetApiRoutes);
+app.use('/api/products', productApiRoutes);
 app.use('/users', usersRoutes);
 // Note: mount other resources here, using the same pattern above
